Revert "Merge branch 'master' of https://github.com/ValoTheValo/CEV-Eris"
This reverts commit 4cf9f09cc122c9a2badc0b741ad5b4815f4806ec, reversing
changes made to f6dba6c9bd22162411581421aa76818dbea559e6.

+++ code/game/objects/items/weapons/weldbackpack.dm
@@ -35,10 +35,17 @@
 		to_chat(user, SPAN_WARNING("The pack is already full!"))
 		return
 
-/obj/item/weldpack/examine(mob/user, extra_description = "")
-	extra_description += "\icon[src] [reagents.total_volume] units of fuel left!"
-	..(user, extra_description)
+/obj/item/weldpack/examine(mob/user)
+	..(user)
+	to_chat(user, text("\icon[] [] units of fuel left!", src, src.reagents.total_volume))
+	return
 
 /obj/item/weldpack/proc/explode()
-	explosion(get_turf(src), reagents.total_volume/2, 50)
-	qdel(src)
+	if (reagents.total_volume > 150)
+		explosion(src.loc,1,2,4)
+	else if (reagents.total_volume > 50)
+		explosion(src.loc,0,1,3)
+	else if (reagents.total_volume > 0)
+		explosion(src.loc,-1,1,2)
+	if(src)
+		qdel(src)
